+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+
+- Added support for Baselibs built with netCDF as a shared library
+- Add `NETCDF_INCLUDE_DIR` as alias to `INC_NETCDF` for spack compatibility
+
 ### Fixed
 
 - Updated the CI to work with latest Baselibs